fix/logging: fix logging by using initializing env_logger when the Client is created
- also applies rustfmt
