+++ src/utils/mod.rs
@@ -181,6 +181,27 @@ pub fn bin_data_format(data: &[u8]) -> String {
     )
 }
 
+/// Initialises `env_logger` with custom settings.
+pub fn init_log() {
+    use std::io::Write;
+    let do_format = move |formatter: &mut env_logger::fmt::Formatter, record: &log::Record<'_>| {
+        let now = formatter.timestamp();
+        writeln!(
+            formatter,
+            "{} {} [{}:{}] {}",
+            formatter.default_styled_level(record.level()),
+            now,
+            record.file().unwrap_or_default(),
+            record.line().unwrap_or_default(),
+            record.args()
+        )
+    };
+    let _ = env_logger::Builder::from_default_env()
+        .format(do_format)
+        .try_init()
+        .unwrap_or(());
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
